requirements.txt
astroid==1.4.4
colorama==0.3.6
Django==1.6.2
django-rest-swagger==0.3.5
djangorestframework==3.2.5
djangorestframework-gis==0.9.6
fake-factory==0.5.3
lazy-object-proxy==1.2.1
psycopg2==2.6.1
pylint==1.5.4
pylint-django==0.7.1
pylint-plugin-utils==0.2.3
PyYAML==3.11
six==1.10.0
wheel==0.24.0
wrapt==1.10.6
